feat(treeset): add treeset api

+++ treeset.ts
@@ -4,34 +4,34 @@ class TreeSet<T = number> {
   #size: number
   tree: RBTree<T>
   compare: (l: T, r: T) => boolean
-  constructor (collection: T[] = [], compare = (l: T, r: T) => l < r) {
+  constructor(collection: T[] = [], compare = (l: T, r: T) => l < r) {
     this.#size = 0
     this.tree = new RBTree(compare)
     this.compare = compare
     for (const val of collection) this.add(val)
   }
 
-  size (): number {
+  size(): number {
     return this.#size
   }
 
-  has (val: T): boolean {
+  has(val: T): boolean {
     return !!this.tree.find(val)
   }
 
-  add (val: T): boolean {
+  add(val: T): boolean {
     const added = this.tree.insert(val)
     this.#size += added ? 1 : 0
     return added
   }
 
-  delete (val: T): boolean {
+  delete(val: T): boolean {
     const deleted = this.tree.deleteByValue(val)
     this.#size -= deleted ? 1 : 0
     return deleted
   }
 
-  ceiling (val: T): T | undefined {
+  ceiling(val: T): T | undefined {
     let p = this.tree.root
     let higher = null
     while (p) {
@@ -45,7 +45,7 @@ class TreeSet<T = number> {
     return higher?.data
   }
 
-  floor (val: T): T | undefined {
+  floor(val: T): T | undefined {
     let p = this.tree.root
     let lower = null
     while (p) {
@@ -59,7 +59,7 @@ class TreeSet<T = number> {
     return lower?.data
   }
 
-  higher (val: T): T | undefined {
+  higher(val: T): T | undefined {
     let p = this.tree.root
     let higher = null
     while (p) {
@@ -73,7 +73,7 @@ class TreeSet<T = number> {
     return higher?.data
   }
 
-  lower (val: T): T | undefined {
+  lower(val: T): T | undefined {
     let p = this.tree.root
     let lower = null
     while (p) {
@@ -87,15 +87,37 @@ class TreeSet<T = number> {
     return lower?.data
   }
 
-  * [Symbol.iterator] (): Generator<T, void, void> {
+  first(): T | undefined {
+    return this.tree.inOrder().next().value
+  }
+
+  last(): T | undefined {
+    return this.tree.reverseInOrder().next().value
+  }
+
+  shift(): T | undefined {
+    const first = this.first()
+    if (first === undefined) return undefined
+    this.delete(first)
+    return first
+  }
+
+  pop(): T | undefined {
+    const last = this.last()
+    if (last === undefined) return undefined
+    this.delete(last)
+    return last
+  }
+
+  *[Symbol.iterator](): Generator<T, void, void> {
     for (const val of this.values()) yield val
   }
 
-  * keys (): Generator<T, void, void> {
+  *keys(): Generator<T, void, void> {
     for (const val of this.values()) yield val
   }
 
-  * values (): Generator<T, undefined, void> {
+  *values(): Generator<T, undefined, void> {
     for (const val of this.tree.inOrder()) yield val
     return undefined
   }
@@ -103,7 +125,7 @@ class TreeSet<T = number> {
   /**
    * Return a generator for reverse order traversing the set
    */
-  * rvalues (): Generator<T, undefined, void> {
+  *rvalues(): Generator<T, undefined, void> {
     for (const val of this.tree.reverseInOrder()) yield val
     return undefined
   }
@@ -114,7 +136,7 @@ class TreeMultiSet<T = number> {
   tree: RBTree<T>
   counts: Map<T, number>
   compare: (l: T, r: T) => boolean
-  constructor (collection: T[] = [], compare = (l: T, r: T) => l < r) {
+  constructor(collection: T[] = [], compare = (l: T, r: T) => l < r) {
     this.#size = 0
     this.tree = new RBTree(compare)
     this.counts = new Map()
@@ -122,33 +144,36 @@ class TreeMultiSet<T = number> {
     for (const val of collection) this.add(val)
   }
 
-  size (): number {
+  size(): number {
     return this.#size
   }
 
-  has (val: T): boolean {
+  has(val: T): boolean {
     return !!this.tree.find(val)
   }
 
-  add (val: T): void {
-    this.tree.insert(val)
+  add(val: T): boolean {
+    const added = this.tree.insert(val)
     this.increase(val)
     this.#size++
+    return added
   }
 
-  delete (val: T): void {
+  delete(val: T): boolean {
+    if (!this.has(val)) return false
     this.decrease(val)
     if (this.count(val) === 0) {
       this.tree.deleteByValue(val)
     }
     this.#size--
+    return true
   }
 
-  count (val: T): number {
+  count(val: T): number {
     return this.counts.get(val) ?? 0
   }
 
-  ceiling (val: T): T | undefined {
+  ceiling(val: T): T | undefined {
     let p = this.tree.root
     let higher = null
     while (p) {
@@ -162,7 +187,7 @@ class TreeMultiSet<T = number> {
     return higher?.data
   }
 
-  floor (val: T): T | undefined {
+  floor(val: T): T | undefined {
     let p = this.tree.root
     let lower = null
     while (p) {
@@ -176,7 +201,7 @@ class TreeMultiSet<T = number> {
     return lower?.data
   }
 
-  higher (val: T): T | undefined {
+  higher(val: T): T | undefined {
     let p = this.tree.root
     let higher = null
     while (p) {
@@ -190,7 +215,7 @@ class TreeMultiSet<T = number> {
     return higher?.data
   }
 
-  lower (val: T): T | undefined {
+  lower(val: T): T | undefined {
     let p = this.tree.root
     let lower = null
     while (p) {
@@ -204,11 +229,37 @@ class TreeMultiSet<T = number> {
     return lower?.data
   }
 
-  * keys (): Generator<T, void, void> {
+  first(): T | undefined {
+    return this.tree.inOrder().next().value
+  }
+
+  last(): T | undefined {
+    return this.tree.reverseInOrder().next().value
+  }
+
+  shift(): T | undefined {
+    const first = this.first()
+    if (first === undefined) return undefined
+    this.delete(first)
+    return first
+  }
+
+  pop(): T | undefined {
+    const last = this.last()
+    if (last === undefined) return undefined
+    this.delete(last)
+    return last
+  }
+
+  *[Symbol.iterator](): Generator<T, void, void> {
+    yield* this.values()
+  }
+
+  *keys(): Generator<T, void, void> {
     for (const val of this.values()) yield val
   }
 
-  * values (): Generator<T, undefined, void> {
+  *values(): Generator<T, undefined, void> {
     for (const val of this.tree.inOrder()) {
       let count = this.count(val)
       while (count--) yield val
@@ -219,7 +270,7 @@ class TreeMultiSet<T = number> {
   /**
    * Return a generator for reverse order traversing the multi-set
    */
-  * rvalues (): Generator<T, undefined, void> {
+  *rvalues(): Generator<T, undefined, void> {
     for (const val of this.tree.reverseInOrder()) {
       let count = this.count(val)
       while (count--) yield val
@@ -230,14 +281,14 @@ class TreeMultiSet<T = number> {
   /**
    * Should only be called by `delete` to keep the internal state correct
    */
-  private decrease (val: T): void {
+  private decrease(val: T): void {
     this.counts.set(val, this.count(val) - 1)
   }
   /**
    * Should only be called by `add` to keep the internal state correct
    */
 
-  private increase (val: T): void {
+  private increase(val: T): void {
     this.counts.set(val, this.count(val) + 1)
   }
 }
